Agri/requirements.txt
Django==1.11.20
djangorestframework==3.9.2
PyMySQL==0.9.3
pytz==2019.1
